TypeIndicator정보 질문 엔티티 대신 답변 엔티티에 담도록 변경, 유형 설명 및 이미지 넘버링 추가

+++ src/main/java/com/typetest/personalities/dto/PersonalitiesAnswerInfo.java
@@ -13,6 +13,6 @@ import java.util.HashMap;
 public class PersonalitiesAnswerInfo {
     private Long userId;
     private AnswerType answerType;
-    private HashMap<Integer, Integer> answer;
+    private HashMap<Integer, Long> answer;
     private TestCodeInfo testCodeInfo;
 }
